Move parts of WeakReference to shared partition (#24800)
